pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = "Tech blog with Django + Tailwind + Markdown, tags, search, RSS"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.2.1"
gunicorn = "^23.0.0"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
django-taggit = "^6.1.0"
django-unfold = "^0.76.0"
django-browser-reload = "^1.21.0"
django-tailwind = "^4.4.0"
whitenoise = "^6.7.0"
markdown = "^3.6"
PyYAML = "^6.0.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
